+++ psPAS/Functions/LDAPDirectories/Set-PASDirectoryMapping.ps1
@@ -62,7 +62,27 @@ function Set-PASDirectoryMapping {
 			ValueFromPipelinebyPropertyName = $true
 		)]
 		[ValidateRange(1, 3650)]
-		[int]$UserActivityLogPeriod
+		[int]$UserActivityLogPeriod,
+
+		[parameter(
+			Mandatory = $false,
+			ValueFromPipelinebyPropertyName = $true
+		)]
+		[int]$UsedQuota,
+
+		[parameter(
+			Mandatory = $false,
+			ValueFromPipelinebyPropertyName = $true
+		)]
+		[ValidateSet('PIMSU', 'PSM', 'PSMP', 'PVWA', 'WINCLIENT', 'PTA', 'PACLI', 'NAPI', 'XAPI', 'HTTPGW',
+			'EVD', 'PIMSu', 'AIMApp', 'CPM', 'PVWAApp', 'PSMApp', 'AppPrv', 'AIMApp', 'PSMPApp', 'GUI')]
+		[string[]]$AuthorizedInterfaces,
+
+		[parameter(
+			Mandatory = $false,
+			ValueFromPipelinebyPropertyName = $true
+		)]
+		[boolean]$EnableENEWhenDisconnected
 
 	)
 
@@ -93,6 +113,14 @@ function Set-PASDirectoryMapping {
 
 				#v10.10
 				Assert-VersionRequirement -RequiredVersion 10.10
+				Continue
+
+			}
+
+			{ $_ -match 'UsedQuota|AuthorizedInterfaces|EnableENEWhenDisconnected' } {
+
+				#v10.7
+				Assert-VersionRequirement -RequiredVersion 14.0
 				Continue
 
 			}